requirements.txt
selenium
beautifulsoup4
pandas
openpyxl
python-dotenv
webdriver-manager